+++ base/src/main/java/eu/fraho/spring/securityJwt/CryptPasswordEncoder.java
@@ -6,12 +6,14 @@
  */
 package eu.fraho.spring.securityJwt;
 
+import eu.fraho.spring.securityJwt.config.CryptConfiguration;
 import eu.fraho.spring.securityJwt.dto.CryptAlgorithm;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.codec.digest.Crypt;
 import org.jetbrains.annotations.NotNull;
-import org.springframework.beans.factory.InitializingBean;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 
@@ -22,19 +24,13 @@ import java.util.Random;
 
 @Component
 @Slf4j
-public class CryptPasswordEncoder implements PasswordEncoder, InitializingBean {
-    public static final String ALGORITHM_DEFAULT = "SHA512";
-
-    public static final int ROUNDS_MIN = 100;
-    public static final int ROUNDS_DEFAULT = 10_000;
-    public static final int ROUNDS_MAX = 500_000;
-
+@RequiredArgsConstructor(onConstructor = @__(@Autowired))
+public class CryptPasswordEncoder implements PasswordEncoder {
     private final Random random = new SecureRandom();
     private final Encoder encoder = Base64.getUrlEncoder();
-    @Value("${fraho.crypt.rounds:" + ROUNDS_DEFAULT + "}")
-    private int rounds = ROUNDS_DEFAULT;
-    @Value("${fraho.crypt.algorithm:" + ALGORITHM_DEFAULT + "}")
-    private CryptAlgorithm algorithm = CryptAlgorithm.valueOf(ALGORITHM_DEFAULT);
+
+    @NonNull
+    private final CryptConfiguration configuration;
 
     private static boolean slowEquals(@NotNull CharSequence a, @NotNull CharSequence b) {
         int diff = a.length() ^ b.length();
@@ -47,8 +43,9 @@ public class CryptPasswordEncoder implements PasswordEncoder, InitializingBean {
     @Override
     public String encode(@NotNull CharSequence rawPassword) {
         final String cryptParam;
+        CryptAlgorithm algorithm = configuration.getAlgorithm();
         if (algorithm.isRoundsSupported()) {
-            cryptParam = String.format("%srounds=%d$%s$", algorithm.getPrefix(), rounds, generateSalt());
+            cryptParam = String.format("%srounds=%d$%s$", algorithm.getPrefix(), configuration.getRounds(), generateSalt());
         } else if (CryptAlgorithm.DES.equals(algorithm)) {
             cryptParam = generateSalt();
         } else {
@@ -67,25 +64,11 @@ public class CryptPasswordEncoder implements PasswordEncoder, InitializingBean {
 
     @NotNull
     protected String generateSalt() {
+        CryptAlgorithm algorithm = configuration.getAlgorithm();
         final byte[] bytes = new byte[algorithm.getSaltLength() * 2];
         random.nextBytes(bytes);
         String salt = encoder.encodeToString(bytes);
         salt = salt.replaceAll("[-_]", "");
         return salt.substring(0, algorithm.getSaltLength());
-    }
-
-    @Override
-    public void afterPropertiesSet() {
-        if (algorithm.isRoundsSupported()) {
-            if (rounds < ROUNDS_MIN || rounds > ROUNDS_MAX) {
-                log.warn("Encryption rounds out of bounds ({} <= {} <= {}), forcing to default ({})",
-                        ROUNDS_MIN, rounds, ROUNDS_MAX, ROUNDS_DEFAULT);
-                rounds = ROUNDS_DEFAULT;
-            }
-        }
-
-        if (algorithm.isInsecure()) {
-            log.warn("Using insecure crypt variant {}. Consider upgrading to a stronger one.", algorithm);
-        }
     }
 }