+++ base/src/main/java/eu/fraho/spring/securityJwt/service/NullTokenStore.java
@@ -16,7 +16,7 @@ import java.util.List;
 import java.util.Map;
 
 @Slf4j
-class NullTokenStore implements RefreshTokenStore {
+public class NullTokenStore implements RefreshTokenStore {
     @Override
     public void saveToken(@NotNull String username, @NotNull String deviceId, @NotNull String token) {
         throw new FeatureNotConfiguredException("No implementation configured");