+++ base/src/main/java/eu/fraho/spring/securityJwt/service/RegisterRefreshTokenStore.java
@@ -6,11 +6,11 @@
  */
 package eu.fraho.spring.securityJwt.service;
 
+import eu.fraho.spring.securityJwt.config.JwtRefreshConfiguration;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
 import org.springframework.beans.factory.support.AbstractBeanDefinition;
 import org.springframework.beans.factory.support.BeanDefinitionRegistry;
@@ -21,23 +21,19 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class RegisterRefreshTokenStore implements InitializingBean {
     public static final String BEAN_NAME = "refreshTokenStore";
-    private static final String DEFAULT_TOKEN_STORE = "eu.fraho.spring.securityJwt.service.NullTokenStore";
 
     @NonNull
     private final ConfigurableListableBeanFactory factory;
 
-    @Value("${fraho.jwt.refresh.cache.impl:" + DEFAULT_TOKEN_STORE + "}")
-    private Class<? extends RefreshTokenStore> refreshTokenStoreImpl = null;
+    @NonNull
+    private final JwtRefreshConfiguration refreshConfig;
 
     @Override
     public void afterPropertiesSet() throws Exception {
-        if (refreshTokenStoreImpl == null) {
-            refreshTokenStoreImpl = NullTokenStore.class;
-        }
         BeanDefinitionRegistry registry = ((BeanDefinitionRegistry) factory);
 
         GenericBeanDefinition beanDefinition = new GenericBeanDefinition();
-        beanDefinition.setBeanClass(refreshTokenStoreImpl);
+        beanDefinition.setBeanClass(refreshConfig.getCacheImpl());
         beanDefinition.setLazyInit(false);
         beanDefinition.setAbstract(false);
         beanDefinition.setAutowireCandidate(true);