+++ base/src/main/java/eu/fraho/spring/securityJwt/service/TotpServiceImpl.java
@@ -6,10 +6,14 @@
  */
 package eu.fraho.spring.securityJwt.service;
 
+import eu.fraho.spring.securityJwt.config.TotpConfiguration;
 import lombok.Getter;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.codec.binary.Base32;
 import org.springframework.beans.factory.InitializingBean;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
@@ -23,25 +27,13 @@ import java.util.Random;
 
 @Component
 @Slf4j
-public class TotpServiceImpl implements TotpService, InitializingBean {
-    public static final int TOTP_LENGTH_MIN = 8;
-    public static final int TOTP_LENGTH_DEFAULT = 16;
-    public static final int TOTP_LENGTH_MAX = 32;
-
-    public static final int TOTP_VARIANCE_MIN = 1;
-    public static final int TOTP_VARIANCE_DEFAULT = 3;
-    public static final int TOTP_VARIANCE_MAX = 10;
-
+@RequiredArgsConstructor(onConstructor = @__(@Autowired))
+public class TotpServiceImpl implements TotpService {
     private final Base32 base32 = new Base32();
     private final Random random = new SecureRandom();
 
-    @Value("${fraho.totp.variance:" + TOTP_VARIANCE_DEFAULT + "}")
-    @Getter
-    private Integer totpVariance = TOTP_VARIANCE_DEFAULT;
-
-    @Value("${fraho.totp.length:" + TOTP_LENGTH_DEFAULT + "}")
-    @Getter
-    private Integer totpLength = TOTP_LENGTH_DEFAULT;
+    @NonNull
+    private final TotpConfiguration configuration;
 
     // TODO remove / make package private
     public long getCurrentCodeForTesting(String secret) throws InvalidKeyException, NoSuchAlgorithmException {
@@ -72,7 +64,7 @@ public class TotpServiceImpl implements TotpService, InitializingBean {
         final byte[] secretBytes = base32.decode(secret);
         boolean result = false;
         try {
-            for (int i = -totpVariance; i <= totpVariance; i++) {
+            for (int i = -configuration.getVariance(); i <= configuration.getVariance(); i++) {
                 if (getCode(secretBytes, timeIndex + i) == code) {
                     result = true;
                     break;
@@ -86,22 +78,8 @@ public class TotpServiceImpl implements TotpService, InitializingBean {
 
     @Override
     public String generateSecret() {
-        final byte[] secret = new byte[totpLength];
+        final byte[] secret = new byte[configuration.getLength()];
         random.nextBytes(secret);
         return base32.encodeToString(secret);
-    }
-
-    @Override
-    public void afterPropertiesSet() {
-        if (totpVariance < TOTP_VARIANCE_MIN || totpVariance > TOTP_VARIANCE_MAX) {
-            log.warn("TOTP variance out of bounds ({} <= {} <= {}), forcing to default ({})",
-                    TOTP_VARIANCE_MIN, totpVariance, TOTP_VARIANCE_MAX, TOTP_VARIANCE_DEFAULT);
-            totpVariance = TOTP_VARIANCE_DEFAULT;
-        }
-        if (totpLength < TOTP_LENGTH_MIN || totpLength > TOTP_LENGTH_MAX) {
-            log.warn("TOTP length out of bounds ({} <= {} <= {}), forcing to default ({})",
-                    TOTP_LENGTH_MIN, totpLength, TOTP_LENGTH_MAX, TOTP_LENGTH_DEFAULT);
-            totpLength = TOTP_LENGTH_DEFAULT;
-        }
     }
 }