+++ base/src/test/java/eu/fraho/spring/securityJwt/tokenService/TestJwtServiceCreateTokenHmac.java
@@ -39,7 +39,7 @@ public class TestJwtServiceCreateTokenHmac extends AbstractCreateTokenTest {
     public void testValidateFixHmac() throws Exception {
         final String token = "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ4IiwibmJmIjoxNDk1MTE4NzE4LCJleHAiOjI0OTUxMTg3MjAsImlhdCI6MTQ5NTExODcxOH0.ARMN2DMX2mybWsvYRxyX_CAZBHU0nhGJ3qGFbn6uj10";
         Files.write(Paths.get(OUT_KEY), STATIC_SECRET.getBytes(StandardCharsets.UTF_8));
-        jwtTokenService.afterPropertiesSet();
+        tokenConfiguration.afterPropertiesSet();
         Assert.assertTrue("Changed hmac key and validate hard coded token", jwtTokenService.validateToken(token));
     }
 
@@ -47,7 +47,7 @@ public class TestJwtServiceCreateTokenHmac extends AbstractCreateTokenTest {
     public void testValidateCheckNbf() throws Exception {
         final String token = "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ4IiwibmJmIjoyNDk1MTE4NzE4LCJleHAiOjI0OTUxMTg3MjAsImlhdCI6MTQ5NTExODcxOH0.h9jUPjrZ8vubfLqQrUYtzjjY-Zmf4kYEE6Jq7mq2T10";
         Files.write(Paths.get(OUT_KEY), STATIC_SECRET.getBytes(StandardCharsets.UTF_8));
-        jwtTokenService.afterPropertiesSet();
+        tokenConfiguration.afterPropertiesSet();
         Assert.assertFalse("Accepted nbf value in the future", jwtTokenService.validateToken(token));
     }
 
@@ -55,7 +55,7 @@ public class TestJwtServiceCreateTokenHmac extends AbstractCreateTokenTest {
     public void testValidateCheckIat() throws Exception {
         final String token = "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ4IiwibmJmIjoxNDk1MTE4NzE4LCJleHAiOjI0OTUxMTg3MjAsImlhdCI6MjQ5NTExODcxOH0.FjQKAuwUqJfDpl4rO4fL-zKZ5P9BvjI7jI_q5EOyYBY";
         Files.write(Paths.get(OUT_KEY), STATIC_SECRET.getBytes(StandardCharsets.UTF_8));
-        jwtTokenService.afterPropertiesSet();
+        tokenConfiguration.afterPropertiesSet();
         Assert.assertFalse("Accepted iat value in the future", jwtTokenService.validateToken(token));
     }
 
@@ -63,7 +63,7 @@ public class TestJwtServiceCreateTokenHmac extends AbstractCreateTokenTest {
     public void testValidateEmptyIat() throws Exception {
         final String token = "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ4IiwibmJmIjoxNDk1MTE4NzE4LCJleHAiOjI0OTUxMTg3MjB9.SIxUTmTLA0fW6nTh11GiLKhxZ8eikJjk-HeChogJnYc";
         Files.write(Paths.get(OUT_KEY), STATIC_SECRET.getBytes(StandardCharsets.UTF_8));
-        jwtTokenService.afterPropertiesSet();
+        tokenConfiguration.afterPropertiesSet();
         Assert.assertTrue("Declined empty iat", jwtTokenService.validateToken(token));
     }
 
@@ -71,7 +71,7 @@ public class TestJwtServiceCreateTokenHmac extends AbstractCreateTokenTest {
     public void testValidateEmptyExp() throws Exception {
         final String token = "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ4IiwibmJmIjoxNDk1MTE4NzE4LCJpYXQiOjE0OTUxMTg3MTh9.gLl5Yss7YP9vVVwPf_XI9CDTlza4VdVdvbI5cuE-sgM";
         Files.write(Paths.get(OUT_KEY), STATIC_SECRET.getBytes(StandardCharsets.UTF_8));
-        jwtTokenService.afterPropertiesSet();
+        tokenConfiguration.afterPropertiesSet();
         Assert.assertFalse("Accepted empty exp", jwtTokenService.validateToken(token));
     }
 
@@ -79,7 +79,7 @@ public class TestJwtServiceCreateTokenHmac extends AbstractCreateTokenTest {
     public void testValidateEmptyNbf() throws Exception {
         final String token = "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ4IiwiZXhwIjoyNDk1MTE4NzIwLCJpYXQiOjE0OTUxMTg3MTh9.Su-LzJZMPgV2n_OlZnpFjrji95CNMzfvCiVOaRPXyJg";
         Files.write(Paths.get(OUT_KEY), STATIC_SECRET.getBytes(StandardCharsets.UTF_8));
-        jwtTokenService.afterPropertiesSet();
+        tokenConfiguration.afterPropertiesSet();
         Assert.assertTrue("Declined empty nbf", jwtTokenService.validateToken(token));
     }
 }